website use blog post models
Only display 4 news articles on homepage
Have a page that list all news articles
