+++ api/templates/website/index.html
@@ -1,34 +1,53 @@
 {% extends "website/new-base.html" %}
 {% block content %}
-    <div class="container">
-        <div class="row">
-            {% for article in news %}
-                <div class="col-sm-6 col-xs-12">
-                    <h4> {{article.title}} </h4>
-                    {% if article.image %}
-                        <img
-                            src="{{ url_for('website.post_picture', name=article.image) }}"
-                            alt="Article Post Image - {{article.title}}"
-                            class="postImage"
-                            height="200"
-                            width="200" />
-                    {% endif %}
-                    <p>
-                        {{article.summary}} 
-                        <a 
-                            data-cy="{{article.title}}"
-                            alt="{{article.title}}"
-                            href="{{ url_for('website.checkout_post', year=year, date=article.date, file_name=article.name) }}"
-                            hx-get="{{ url_for('website.checkout_post', year=year, date=article.date, file_name=article.name) }}"
-                            hx-target="#website-body"
-                            hx-push-url="true"
-                        >
-                            Read more
-                        </a>
-                    </p>
-                </div>
-            {% endfor %}
+    <div class="news">
+        {% for post in posts %}
+        <div class="article">
+            <h4 class="article__title">
+                {{post.title}}
+            </h4>
+            {% if post.image %}
+                <img
+                    src="{{post.image.url}}"
+                    alt="Article Post Image - {{post.title}}"
+                    class="article__image"
+                    height="200"
+                    width="200" />
+            {% endif %}
+            <p>
+                <strong>Date:</strong> {{post.date}}
+            </p>
+            <p class="article__summary">
+                {{post.summary|safe}} 
+            </p>
+            <a 
+                data-cy="{{post.title}}"
+                class="btn btn-link"
+                alt="{{post.title}}"
+                href="{{ url_for('website.blog_post', year=year, blog_post_id=post.blog_post_id) }}"
+                hx-get="{{ url_for('website.blog_post', year=year, blog_post_id=post.blog_post_id) }}"
+                hx-swap="innerHTML show:top"
+                hx-target="#website-body"
+                hx-push-url="true"
+            >
+                Read more
+            </a>
         </div>
+        {% endfor %}
     </div>
-
+    {% if show_more %}
+        <center>
+            <a 
+                class="btn btn-secondary btn-lg"
+                alt="More news"
+                href="{{ url_for('website.blog_posts', year=year) }}"
+                hx-get="{{ url_for('website.blog_posts', year=year) }}"
+                hx-swap="innerHTML show:top"
+                hx-target="#website-body"
+                hx-push-url="true"
+            >
+                More news
+            </a>
+        </center>
+    {% endif %}
 {% endblock %}
